+++ app/(home)/army-builder/_layout.tsx
@@ -211,6 +211,7 @@ const ArmyBuilderLayout = () => {
     newArmy,
     routeName,
     unitSelectionButtonDisabled,
+    units,
     warlord
   ])